refactor : checkstyle 적용
