+++ pickly-service/src/main/java/org/pickly/service/member/common/MemberMapper.java
@@ -4,10 +4,10 @@ import org.pickly.service.member.controller.request.MemberProfileUpdateReq;
 import org.pickly.service.member.controller.response.MemberModeRes;
 import org.pickly.service.member.controller.response.MemberProfileRes;
 import org.pickly.service.member.entity.Member;
+import org.pickly.service.member.notification.entity.NotificationStandard;
 import org.pickly.service.member.service.dto.MemberModeDTO;
 import org.pickly.service.member.service.dto.MemberProfileDTO;
 import org.pickly.service.member.service.dto.MemberProfileUpdateDTO;
-import org.pickly.service.notification.entity.NotificationStandard;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -21,6 +21,10 @@ public class MemberMapper {
         .profileEmoji(dto.getProfileEmoji())
         .isFollowing(dto.getIsFollowing())
         .build();
+  }
+
+  public MemberModeRes toResponse(MemberModeDTO dto) {
+    return new MemberModeRes(dto.getIsHardMode(), dto.getStandardDate());
   }
 
   public MemberProfileUpdateDTO toDTO(MemberProfileUpdateReq request) {
@@ -44,10 +48,6 @@ public class MemberMapper {
         .isHardMode(member.getIsHardMode())
         .standardDate(standard.getStandardDate())
         .build();
-  }
-
-  public MemberModeRes toResponse(MemberModeDTO dto) {
-    return new MemberModeRes(dto.getIsHardMode(), dto.getStandardDate());
   }
 
 }
