+++ pickly-service/src/main/java/org/pickly/service/member/notification/repository/interfaces/NotificationStandardRepository.java
@@ -1,7 +1,7 @@
-package org.pickly.service.notification.repository.interfaces;
+package org.pickly.service.member.notification.repository.interfaces;
 
 import java.util.Optional;
-import org.pickly.service.notification.entity.NotificationStandard;
+import org.pickly.service.member.notification.entity.NotificationStandard;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface NotificationStandardRepository extends JpaRepository<NotificationStandard, Long> {
